Cleanup
Moved stuff around.
