+++ src/util/arrayops.h
@@ -7,4 +7,13 @@ void swap(int *data, const unsigned int index1, const unsigned int index2) {
     data[index2] = tmp;
 }
 
+int is_sorted(const unsigned int length, const int *data) {
+    for (unsigned int i = 1; i < length; i++) {
+        if(data[i-1] > data[i]) {
+            return 0;
+        }
+    }
+    return 1;
+}
+
 #endif
