+++ src/util/leibnizlib.h
@@ -34,11 +34,3 @@ int readDataset(signed int *array, int LIMIT) {
     return EXIT_SUCCESS;
 }
 
-int is_sorted(const unsigned int length, const int *data) {
-    for (unsigned int i = 1; i < length; i++) {
-        if(data[i-1] > data[i]) {
-            return 0;
-        }
-    }
-    return 1;
-}
